requirements.txt
arabic-reshaper==2.1.3
asgiref==3.4.1
beautifulsoup4==4.9.3
Django==3.2.5
django-bootstrap3==15.0.0
django-bootstrap3-datetimepicker==2.2.3
django-bootstrap4==3.0.1
django-crispy-forms==1.12.0
django-formtools==2.3
django-widget-tweaks==1.4.8
fpdf==1.7.2
fpdf2==2.4.5
future==0.18.2
html5lib==1.1
Pillow==8.3.1
psycopg2-binary==2.9.1
PyPDF2==1.26.0
python-bidi==0.4.2
pytz==2021.1
reportlab==3.6.1
six==1.16.0
soupsieve==2.2.1
sqlparse==0.4.1
webencodings==0.5.1
xhtml2pdf==0.2.5
